+++ ark/schema/shared/declare.ts
@@ -10,6 +10,7 @@ type withMetaPrefixedKeys<o> = {
 
 export interface BaseMeta extends JsonSchema.Meta {
 	alias?: string
+	optional?: true
 }
 
 declare global {
