+++ ark/type/methods/base.ts
