+++ ark/type/methods/object.ts
